♻ Cleanup Render and Stories
