+++ components/Asset/index.jsx
@@ -1,4 +1,4 @@
-export * from './Asset.jsx'
+export { default as AssetInfo } from './Asset.jsx'
 export { default as Chart } from './Chart'
 export { default as OrderBook } from './OrderBook'
 export { default as TradeHistory } from './TradeHistory'
