+++ components/Nav/SearchSidebar/SearchTable.jsx
@@ -1,4 +1,9 @@
-import { AssetId, AssetName, AssetNameBlock, NameVerifiedWrapper } from 'components/Asset'
+import {
+  AssetId,
+  AssetName,
+  AssetNameBlock,
+  NameVerifiedWrapper
+} from 'components/Asset/Typography'
 import { BodyCopySm, BodyCopyTiny } from 'components/Typography'
 import { mdiCheckDecagram, mdiStar } from '@mdi/js'
 import { useMemo, useCallback } from 'react'
@@ -65,12 +70,7 @@ export const mapToSearchResults = ({
 const TableWrapper = styled.div`
   // position: absolute;
   // inset: 0;
-  overflow-y: auto;
-  scrollbar-width: none;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
+  // overflow-y: auto;
 `
 
 const PairSlash = styled.span`
