+++ components/Tabs/index.jsx
@@ -1,6 +1,6 @@
-import Assets from 'components/Wallet/AssetsTable'
-import OpenOrders from 'components/Wallet/OpenOrdersTable'
-import OrderHistory from 'components/Wallet/TradeHistoryTable'
+import Assets from 'components/Wallet/Table/AssetsTable'
+import OpenOrders from 'components/Wallet/Table/OpenOrdersTable'
+import OrderHistory from 'components/Wallet/Table/TradeHistoryTable'
 import PropTypes from 'prop-types'
 import { useState } from 'react'
 import useTranslation from 'next-translate/useTranslation'
