--- components/Wallet/AssetsTable/index.jsx
@@ -1,2 +1,0 @@
-import WalletAssetsTable from './WalletAssetsTable'
-export default WalletAssetsTable
