--- components/Wallet/OpenOrdersTable/index.jsx
@@ -1,2 +1,0 @@
-import WalletOpenOrdersTable from './WalletOpenOrdersTable'
-export default WalletOpenOrdersTable
