+++ components/Wallet/Table/AssetsTable.stories.js
@@ -1,6 +1,7 @@
 import React from 'react'
-import { AssetsTab as Component, default as ComponentWithData } from './WalletAssetsTable'
+import { AssetsTable as Component, default as ComponentWithData } from './AssetsTable'
 import styled from 'styled-components'
+import { ReactQueryDevtools } from 'react-query/devtools'
 
 const Container = styled.div`
   width: 100vw;
@@ -14,13 +15,14 @@ const Container = styled.div`
 `
 
 export default {
-  title: 'Wallet/Wallet Assets Tab',
+  title: 'Wallet/Table',
   component: Component,
   parameters: { layout: 'fullscreen' },
   decorators: [
     (Story) => (
       <Container>
         <Story />
+        <ReactQueryDevtools initialIsOpen={false} />
       </Container>
     )
   ]
@@ -29,15 +31,15 @@ export default {
 const Template = (args) => <Component {...args} />
 const TemplateWithData = (args) => <ComponentWithData {...args} />
 
-export const AssetsTab = Template.bind({})
-AssetsTab.args = {
+export const AssetsTable = Template.bind({})
+AssetsTable.args = {
   wallet: {
     address: 'TJFFNUYWHPPIYDE4DGGYPGHWKGAPJEWP3DGE5THZS3B2M2XIAPQ2WY3X4I'
   }
 }
 
-export const AssetsTabPreview = TemplateWithData.bind({})
-AssetsTabPreview.args = {
+export const AssetsTablePreview = TemplateWithData.bind({})
+AssetsTablePreview.args = {
   wallet: {
     address: 'TJFFNUYWHPPIYDE4DGGYPGHWKGAPJEWP3DGE5THZS3B2M2XIAPQ2WY3X4I'
   }
