+++ components/Wallet/Table/TradeHistoryTable.stories.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import { PlaceOrderForm as Component } from './PlaceOrderForm'
+import { TradeHistoryTable as Component, default as ComponentWithData } from './TradeHistoryTable'
 import styled from 'styled-components'
-import asset from 'spec/Asset'
+import { ReactQueryDevtools } from 'react-query/devtools'
 
 const Container = styled.div`
   width: 100vw;
   height: 100vh;
-  background: ${({ theme }) => theme.colors.gray['800']};
+  background: ${({ theme }) => theme.colors.gray['700']};
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
@@ -14,13 +14,14 @@ const Container = styled.div`
   margin: 0;
 `
 export default {
-  title: 'Wallet',
+  title: 'Wallet/Table',
   component: Component,
   parameters: { layout: 'fullscreen' },
   decorators: [
     (Story) => (
       <Container>
         <Story />
+        <ReactQueryDevtools initialIsOpen={false} />
       </Container>
     )
   ]
@@ -29,17 +30,16 @@ export default {
 const Template = (args) => <Component {...args} />
 const TemplateWithData = (args) => <ComponentWithData {...args} />
 
-export const PlaceOrderForm = Template.bind({})
-PlaceOrderForm.args = {
-  asset,
+export const TradeHistoryTable = Template.bind({})
+TradeHistoryTable.args = {
   wallet: {
     address: 'TJFFNUYWHPPIYDE4DGGYPGHWKGAPJEWP3DGE5THZS3B2M2XIAPQ2WY3X4I'
   }
 }
 
-// export const OpenOrdersPreview = TemplateWithData.bind({})
-// OpenOrdersPreview.args = {
-//     wallet: {
-//         address: 'TJFFNUYWHPPIYDE4DGGYPGHWKGAPJEWP3DGE5THZS3B2M2XIAPQ2WY3X4I'
-//     }
-// }
+export const TradeHistoryTablePreview = TemplateWithData.bind({})
+TradeHistoryTablePreview.args = {
+  wallet: {
+    address: 'TJFFNUYWHPPIYDE4DGGYPGHWKGAPJEWP3DGE5THZS3B2M2XIAPQ2WY3X4I'
+  }
+}
