--- components/Wallet/TradeHistoryTable/index.jsx
@@ -1,2 +1,0 @@
-import WalletTradeHistoryTable from './WalletTradeHistoryTable'
-export default WalletTradeHistoryTable
