+++ components/Wallet/WalletTabs.jsx
@@ -1,7 +1,7 @@
 import useTranslation from 'next-translate/useTranslation'
-import { default as WalletOpenOrdersTable } from './OpenOrdersTable/WalletOpenOrdersTable'
-import { default as WalletTradeHistoryTable } from './TradeHistoryTable/WalletTradeHistoryTable'
-import { default as WalletAssetsTable } from './AssetsTable/WalletAssetsTable'
+import { default as WalletOpenOrdersTable } from './Table/OpenOrdersTable'
+import { default as WalletTradeHistoryTable } from './Table/TradeHistoryTable'
+import { default as WalletAssetsTable } from './Table/AssetsTable'
 import { useState } from 'react'
 import { Container, Header, Tab } from 'components/Tabs'
 import PropTypes from 'prop-types'
